+++ packages/combobox/src/Combobox.spec.tsx
@@ -167,6 +167,7 @@ describe('packages/combobox', () => {
           {
             value: 'paragraph',
             displayName,
+            isDisabled: false,
           },
         ];
 
@@ -174,6 +175,70 @@ describe('packages/combobox', () => {
         const { optionElements } = openMenu();
         const [optionEl] = Array.from(optionElements!);
         expect(optionEl).toHaveTextContent(displayName);
+      });
+
+      test('Disabled options are still rendered in the menu', () => {
+        const options: Array<OptionObject> = [
+          ...defaultOptions,
+          {
+            value: 'disabled',
+            displayName: 'Disabled',
+            isDisabled: true,
+          },
+        ];
+        const { openMenu } = renderCombobox(select, { options });
+        const { optionElements } = openMenu();
+        expect(optionElements).toHaveLength(defaultOptions.length + 1);
+      });
+
+      test('Disabled option is not selectable with the mouse', () => {
+        const options: Array<OptionObject> = [
+          {
+            value: 'disabled',
+            displayName: 'Disabled',
+            isDisabled: true,
+          },
+          ...defaultOptions,
+        ];
+        const initialValue = select === 'multiple' ? ['apple'] : 'apple';
+        const { openMenu, inputEl, queryChipsByName } = renderCombobox(select, {
+          options,
+          initialValue,
+        });
+        const { optionElements } = openMenu();
+        const disabledOption = optionElements?.[0];
+        userEvent.click(disabledOption as HTMLLIElement);
+        if (select === 'multiple') {
+          expect(queryChipsByName('Apple')).toBeInTheDocument();
+          expect(queryChipsByName('Disabled')).not.toBeInTheDocument();
+        } else {
+          expect(inputEl).toHaveValue('Apple');
+        }
+      });
+
+      test('Disabled option is not selectable with the keyboard', () => {
+        const options: Array<OptionObject> = [
+          {
+            value: 'disabled',
+            displayName: 'Disabled',
+            isDisabled: true,
+          },
+          ...defaultOptions,
+        ];
+        const initialValue = select === 'multiple' ? ['apple'] : 'apple';
+        const { openMenu, inputEl, queryChipsByName } = renderCombobox(select, {
+          options,
+          initialValue,
+        });
+        const { optionElements } = openMenu();
+        const disabledOption = optionElements![0];
+        userEvent.type(disabledOption, `{enter}`);
+        if (select === 'multiple') {
+          expect(queryChipsByName('Apple')).toBeInTheDocument();
+          expect(queryChipsByName('Disabled')).not.toBeInTheDocument();
+        } else {
+          expect(inputEl).toHaveValue('Apple');
+        }
       });
 
       // Grouped Options
@@ -243,6 +308,7 @@ describe('packages/combobox', () => {
             {
               value: 'paragraph',
               displayName,
+              isDisabled: false,
             },
             ...defaultOptions,
           ];
@@ -480,7 +546,7 @@ describe('packages/combobox', () => {
       );
 
       testSingleSelect(
-        'Input returned to previous selection when menu closes',
+        'Input returned to previous valid selection when menu closes',
         () => {
           const initialValue = 'apple';
           const { inputEl } = renderCombobox(select, {
@@ -492,7 +558,7 @@ describe('packages/combobox', () => {
       );
 
       testSingleSelect(
-        'Clicking the combobox after making a selection should re-open the menu',
+        'Clicking after making a selection should re-open the menu',
         async () => {
           const { comboboxEl, inputEl, openMenu, getMenuElements } =
             renderCombobox(select);
@@ -501,7 +567,7 @@ describe('packages/combobox', () => {
           userEvent.click(firstOption);
           await waitForElementToBeRemoved(menuContainerEl);
           userEvent.click(comboboxEl);
-          await waitFor(() => {
+          waitFor(() => {
             const { menuContainerEl: newMenuContainerEl } = getMenuElements();
             expect(newMenuContainerEl).not.toBeNull();
             expect(newMenuContainerEl).toBeInTheDocument();
@@ -509,6 +575,17 @@ describe('packages/combobox', () => {
           });
         },
       );
+
+      test('Opening the menu when there is a selection should show all options', () => {
+        // See also: 'Pressing Down Arrow when there is a selection shows all menu options'
+        const initialValue = select === 'multiple' ? ['apple'] : 'apple';
+        const { comboboxEl, getMenuElements } = renderCombobox(select, {
+          initialValue,
+        });
+        userEvent.click(comboboxEl);
+        const { optionElements } = getMenuElements();
+        expect(optionElements).toHaveLength(defaultOptions.length);
+      });
 
       describe('Clickaway', () => {
         test('Menu closes on click-away', async () => {
@@ -575,44 +652,43 @@ describe('packages/combobox', () => {
         });
       });
 
-      /**
-       * Clicking buttons
-       */
-      test('Clicking clear all button clears selection', () => {
-        const initialValue =
-          select === 'single' ? 'apple' : ['apple', 'banana', 'carrot'];
-        const { inputEl, clearButtonEl, queryAllChips } = renderCombobox(
-          select,
-          {
-            initialValue,
-          },
-        );
-        expect(clearButtonEl).not.toBeNull();
-        userEvent.click(clearButtonEl!);
-        if (select === 'multiple') {
-          expect(queryAllChips()).toHaveLength(0);
-        } else {
-          expect(inputEl).toHaveValue('');
-        }
-      });
+      describe('Click clear button', () => {
+        test('Clicking clear all button clears selection', () => {
+          const initialValue =
+            select === 'single' ? 'apple' : ['apple', 'banana', 'carrot'];
+          const { inputEl, clearButtonEl, queryAllChips } = renderCombobox(
+            select,
+            {
+              initialValue,
+            },
+          );
+          expect(clearButtonEl).not.toBeNull();
+          userEvent.click(clearButtonEl!);
+          if (select === 'multiple') {
+            expect(queryAllChips()).toHaveLength(0);
+          } else {
+            expect(inputEl).toHaveValue('');
+          }
+        });
 
-      test('Clicking clear all button does nothing when disabled', () => {
-        const initialValue =
-          select === 'single' ? 'apple' : ['apple', 'banana', 'carrot'];
-        const { inputEl, clearButtonEl, queryAllChips } = renderCombobox(
-          select,
-          {
-            initialValue,
-            disabled: true,
-          },
-        );
-        expect(clearButtonEl).not.toBeNull();
-        userEvent.click(clearButtonEl!);
-        if (select === 'multiple') {
-          expect(queryAllChips()).toHaveLength(initialValue.length);
-        } else {
-          expect(inputEl).toHaveValue(startCase(initialValue as string));
-        }
+        test('Clicking clear all button does nothing when disabled', () => {
+          const initialValue =
+            select === 'single' ? 'apple' : ['apple', 'banana', 'carrot'];
+          const { inputEl, clearButtonEl, queryAllChips } = renderCombobox(
+            select,
+            {
+              initialValue,
+              disabled: true,
+            },
+          );
+          expect(clearButtonEl).not.toBeNull();
+          userEvent.click(clearButtonEl!);
+          if (select === 'multiple') {
+            expect(queryAllChips()).toHaveLength(initialValue.length);
+          } else {
+            expect(inputEl).toHaveValue(startCase(initialValue as string));
+          }
+        });
       });
 
       describe('Clicking chips', () => {
@@ -956,6 +1032,32 @@ describe('packages/combobox', () => {
           const reOpenedMenu = await findByRole('listbox');
           expect(reOpenedMenu).toBeInTheDocument();
         });
+
+        test('Pressing Down Arrow when there is a selection shows all menu options', () => {
+          // See also: 'Opening the menu when there is a selection should show all options'
+          const initialValue = select === 'multiple' ? ['apple'] : 'apple';
+          const { inputEl, getMenuElements } = renderCombobox(select, {
+            initialValue,
+          });
+          // First pressing escape to ensure the menu is closed
+          userEvent.type(inputEl, '{esc}{arrowdown}');
+          const { optionElements } = getMenuElements();
+          expect(optionElements).toHaveLength(defaultOptions.length);
+          expect(optionElements![0]).toHaveAttribute('aria-selected', 'true');
+        });
+
+        test('Pressing Up Arrow when there is a selection shows all menu options', () => {
+          // See also: 'Opening the menu when there is a selection should show all options'
+          const initialValue = select === 'multiple' ? ['apple'] : 'apple';
+          const { inputEl, getMenuElements } = renderCombobox(select, {
+            initialValue,
+          });
+          // First pressing escape to ensure the menu is closed
+          userEvent.type(inputEl, '{esc}{arrowup}');
+          const { optionElements } = getMenuElements();
+          expect(optionElements).toHaveLength(defaultOptions.length);
+          expect(optionElements![0]).toHaveAttribute('aria-selected', 'true');
+        });
       });
 
       describe('Left arrow key', () => {
@@ -1139,6 +1241,14 @@ describe('packages/combobox', () => {
             });
           },
         );
+
+        test('Filters the menu options', () => {
+          // Using default options
+          const { inputEl, getMenuElements } = renderCombobox(select);
+          userEvent.type(inputEl, 'c');
+          const { optionElements } = getMenuElements();
+          expect(optionElements).toHaveLength(1); // carrot
+        });
       });
     });
 
@@ -1306,15 +1416,6 @@ describe('packages/combobox', () => {
         );
         expect(errorStateTextEl).toBeInTheDocument();
       });
-    });
-
-    // Filtering
-    test('Menu options list narrows when text is entered', async () => {
-      const { inputEl, openMenu, findAllByRole } = renderCombobox(select);
-      openMenu();
-      userEvent.type(inputEl, 'c');
-      const optionElements = await findAllByRole('option');
-      expect(optionElements.length).toEqual(1);
     });
   });