+++ packages/combobox/src/Combobox.types.ts
@@ -5,12 +5,6 @@ import { Either } from '@leafygreen-ui/lib';
  * Prop Enums & Types
  */
 
-export interface OptionObject {
-  value: string;
-  displayName: string;
-  hasGlyph?: boolean;
-}
-
 export const ComboboxElement = {
   Input: 'Input',
   ClearButton: 'ClearButton',
@@ -28,7 +22,11 @@ export type ComboboxElement =
  */
 
 export const ComboboxSize = {
-  default: 'default',
+  // TODO: add XSmall & Small variants after the refresh
+  // XSmall: 'xsmall',
+  // Small: 'small',
+  Default: 'default',
+  Large: 'large',
 } as const;
 export type ComboboxSize = typeof ComboboxSize[keyof typeof ComboboxSize];
 
@@ -273,6 +271,7 @@ export type ComboboxProps<M extends boolean> = Either<
 /**
  * Combobox Option Props
  */
+
 interface BaseComboboxOptionProps {
   /**
    * The internal value of the option. Used as the identifier in Combobox `initialValue`, value and filteredOptions.
@@ -292,6 +291,12 @@ interface BaseComboboxOptionProps {
   glyph?: ReactElement;
 
   /**
+   * Defines whether the option is disabled.
+   * Node: disabled options are still rendered in the menu, but not selectable.
+   */
+  disabled?: boolean;
+
+  /**
    * Styling Prop
    */
   className?: string;
@@ -302,12 +307,20 @@ export type ComboboxOptionProps = Either<
   'value' | 'displayName'
 >;
 
+export interface OptionObject {
+  value: string;
+  displayName: string;
+  isDisabled: boolean;
+  hasGlyph?: boolean;
+}
+
 export interface InternalComboboxOptionProps {
   value: string;
   displayName: string;
   isSelected: boolean;
   isFocused: boolean;
   setSelected: () => void;
+  disabled?: boolean;
   glyph?: ReactElement;
   className?: string;
   index: number;