+++ packages/combobox/src/ComboboxOption.tsx
@@ -11,12 +11,13 @@ import {
 } from './Combobox.types';
 import { ComboboxContext } from './ComboboxContext';
 import { wrapJSX } from './utils';
+import { fontFamilies } from '@leafygreen-ui/tokens';
 
 /**
  * Styles
  */
 
-const comboboxOptionStyle = () => css`
+const comboboxOptionStyle = css`
   position: relative;
   display: flex;
   align-items: center;
@@ -25,10 +26,12 @@ const comboboxOptionStyle = () => css`
   color: inherit;
   cursor: pointer;
   overflow: hidden;
+  font-family: ${fontFamilies.legacy};
   font-size: var(--lg-combobox-item-font-size);
   line-height: var(--lg-combobox-item-line-height);
   padding: var(--lg-combobox-item-padding-y) var(--lg-combobox-item-padding-x);
 
+  // Left wedge
   &:before {
     content: '';
     position: absolute;
@@ -58,11 +61,21 @@ const comboboxOptionStyle = () => css`
   }
 `;
 
+const comboboxOptionDisabledStyle = css`
+  cursor: not-allowed;
+  color: ${uiColors.gray.base};
+
+  &:hover {
+    background-color: inherit;
+  }
+`;
+
 const flexSpan = css`
   display: inline-flex;
   gap: 8px;
   justify-content: start;
   align-items: inherit;
+  overflow-wrap: anywhere;
 `;
 
 const disallowPointer = css`
@@ -72,10 +85,10 @@ const disallowPointer = css`
 const displayNameStyle = (isSelected: boolean) => css`
   font-weight: ${isSelected ? 'bold' : 'normal'};
 `;
+
 /**
  * Component
  */
-
 const InternalComboboxOption = React.forwardRef<
   HTMLLIElement,
   InternalComboboxOptionProps
@@ -86,6 +99,7 @@ const InternalComboboxOption = React.forwardRef<
       glyph,
       isSelected,
       isFocused,
+      disabled,
       setSelected,
       className,
     }: InternalComboboxOptionProps,
@@ -98,18 +112,22 @@ const InternalComboboxOption = React.forwardRef<
 
     const handleOptionClick = useCallback(
       (e: React.SyntheticEvent) => {
+        // stopPropagation will not stop the keyDown event (only click)
+        // since the option is never `focused`, only `aria-selected`
+        // the keyDown event does not actually fire on the option element
         e.stopPropagation();
 
         // If user clicked on the option, or the checkbox
         // Ignore extra click events on the checkbox wrapper
         if (
-          e.target === optionRef.current ||
-          (e.target as Element).tagName === 'INPUT'
+          !disabled &&
+          (e.target === optionRef.current ||
+            (e.target as Element).tagName === 'INPUT')
         ) {
           setSelected();
         }
       },
-      [optionRef, setSelected],
+      [disabled, optionRef, setSelected],
     );
 
     const renderedIcon = useMemo(() => {
@@ -135,6 +153,7 @@ const InternalComboboxOption = React.forwardRef<
             checked={isSelected}
             tabIndex={-1}
             animate={false}
+            disabled={disabled}
           />
         );
 
@@ -176,6 +195,7 @@ const InternalComboboxOption = React.forwardRef<
       inputValue,
       darkMode,
       optionTextId,
+      disabled,
       withIcons,
     ]);
 
@@ -186,9 +206,15 @@ const InternalComboboxOption = React.forwardRef<
         aria-selected={isFocused}
         aria-label={displayName}
         tabIndex={-1}
-        className={cx(comboboxOptionStyle(), className)}
+        className={cx(
+          comboboxOptionStyle,
+          {
+            [comboboxOptionDisabledStyle]: disabled,
+          },
+          className,
+        )}
         onClick={handleOptionClick}
-        onKeyPress={handleOptionClick}
+        onKeyDown={handleOptionClick}
       >
         {renderedChildren}
       </li>