+++ packages/pipeline/src/Pipeline.tsx
@@ -117,7 +117,6 @@ const getStatefulStyles = (state: StateForStyles) => ({
 });
 
 /**
- * # Pipeline
  *
  * React Component to render top-level MongoDB Aggregations in a visual format.
  * Handles overflowed content gracefully by indicating how many other stages are hidden.
@@ -129,6 +128,7 @@ const getStatefulStyles = (state: StateForStyles) => ({
  *   <Stage>$limit</Stage>
  * </Pipeline>
  * ```
+ *
  * @param props.children Content that will appear inside of the Pipeline component.
  * @param props.className Classname applied to Pipeline content container.
  * @param props.size Alters the rendered size of the component.